Refactor utility assignment scripts
