+++ context/code/data/ny_utility_assignment_resstock.md
@@ -8,10 +8,11 @@ How electric and gas utilities are assigned to ResStock buildings in NY: `data/r
 
 ## Overview
 
-- **Entrypoint:** `assign_utility_ny()` (and CLI via `assign_utility_ny.py`).
+- **Entrypoint:** `assign_utility_ny()` in `data/resstock/utility/assign_utility_ny.py` (and CLI via `assign_utility_ny.py`). This is a thin wrapper that builds the NY utility-name crosswalk and passes NY-specific configuration (utility name map, excluded gas utilities, state CRS) to the generic `create_hh_utilities()` in `data/resstock/utility/utils.py`.
 - **Inputs:** ResStock metadata (with `in.puma`, `in.heating_fuel`, `has_natgas_connection`), electric and gas utility service-territory polygons (CSV with WKT), Census PUMAs (pygris).
 - **Outputs:** Same metadata with `sb.electric_utility` and `sb.gas_utility` added (or overwritten).
 - **Logic:** PUMA–utility overlap → PUMA-level probability tables → per-building sampling (deterministic seed). Electric: every building gets an electric utility. Gas: only buildings with `has_natgas_connection` get a gas utility; others get null.
+- **Generic functions:** `create_hh_utilities()`, `zero_excluded_gas_utilities_and_renormalize()`, `calculate_puma_utility_overlap()`, `calculate_utility_probabilities()`, `calculate_prior_distributions()`, `sample_utility_per_building()`, `print_comparison_summary()`, `puma_id_series_for_join()`, `read_csv_to_gdf_from_s3()` all live in `data/resstock/utility/utils.py` and are state-generic.
 
 ---
 
@@ -22,11 +23,11 @@ A fixed set of **excluded gas utilities** are excluded from assignment: their pr
 - **Constant:** `EXCLUDED_GAS_UTILITIES` in `assign_utility_ny.py` (loaded from `data/resstock/state_configs.yaml` → `NY.excluded_gas_utilities`):
   - `bath`, `chautauqua`, `corning`, `fillmore`, `reserve`, `stlaw`
 - **Rationale:** These utilities have very few customers; we do not assign ResStock buildings to them for rate-design/BAT purposes.
-- **Implementation:** `_zero_small_gas_utilities_and_renormalize()`:
-  1. Set to 0 the probability columns whose name is in `EXCLUDED_GAS_UTILITIES`.
+- **Implementation:** `zero_excluded_gas_utilities_and_renormalize()` (in `data/resstock/utility/utils.py`), called from `create_hh_utilities()` when `excluded_gas_utilities` is non-empty:
+  1. Set to 0 the probability columns whose name is in `excluded_utilities`.
   2. For each PUMA row, if the remaining (non-excluded) gas probabilities sum to zero, the PUMA is "bad" and must be handled (see below).
   3. Otherwise, renormalize each row so gas probabilities sum to 1.
-  4. Final gas probability table is used by `_sample_utility_per_building(..., only_when_fuel="Natural Gas")`.
+  4. Final gas probability table is used by `sample_utility_per_building(..., only_when_fuel="Natural Gas")`.
 
 ---
 
@@ -34,7 +35,7 @@ A fixed set of **excluded gas utilities** are excluded from assignment: their pr
 
 If, for a given PUMA, **all** gas probability was in excluded utilities, then after zeroing that PUMA has no gas utility left. Two behaviors:
 
-1. **`pumas` not provided:** `_zero_small_gas_utilities_and_renormalize(..., pumas=None)` raises `ValueError` with the affected `puma_id`(s).
+1. **`pumas` not provided:** `zero_excluded_gas_utilities_and_renormalize(..., pumas=None)` raises `ValueError` with the affected `puma_id`(s).
 2. **`pumas` provided (GeoDataFrame):** A **donor** PUMA is chosen and its gas probability row is used for the bad PUMA.
    - **Donor selection:** Prefer a **good** PUMA (non-zero gas probability after exclusion) that is **adjacent** to the bad PUMA (geometries touch). Among adjacent good PUMAs, choose the one whose centroid is closest to the bad PUMA's centroid. If no adjacent good PUMA exists, use the good PUMA with the nearest centroid (fallback).
    - **Result:** The bad PUMA's row is replaced by the donor's gas probability row; then all rows are renormalized so each sums to 1.
@@ -46,11 +47,11 @@ If, for a given PUMA, **all** gas probability was in excluded utilities, then af
 
 PUMA identifiers can appear as integers or strings (e.g. `100` vs `"00100"`). For consistent matching between the gas-probability table and the PUMAs GeoDataFrame:
 
-- **`_puma_id_series_for_join(pumas)`** returns a pandas Series of 5-character zero-padded PUMA ids derived from `pumas`:
+- **`puma_id_series_for_join(pumas)`** (in `data/resstock/utility/utils.py`) returns a pandas Series of 5-character zero-padded PUMA ids derived from `pumas`:
   - If `PUMACE10` exists: `pumas["PUMACE10"].astype(str).str.zfill(5)`.
   - Else if `GEOID` exists: last 5 characters of `GEOID`.
   - Else `None`.
-- Bad/donor PUMA matching in `_zero_small_gas_utilities_and_renormalize` uses this normalization (e.g. `str(bad_puma_id).zfill(5)`) so that geometry lookups and probability row replacement are consistent.
+- Bad/donor PUMA matching in `zero_excluded_gas_utilities_and_renormalize` uses this normalization (e.g. `str(bad_puma_id).zfill(5)`) so that geometry lookups and probability row replacement are consistent.
 
 ---
 
@@ -78,6 +79,6 @@ After zeroing excluded gas utilities (and optionally replacing bad-PUMA rows wit
 `tests/test_assign_utility_ny.py` covers:
 
 - `EXCLUDED_GAS_UTILITIES` constant (loaded from `state_configs.yaml`).
-- `_puma_id_series_for_join` (PUMACE10, GEOID, missing columns).
-- `_zero_small_gas_utilities_and_renormalize`: no excluded cols (unchanged); zero + renormalize; bad PUMA with `pumas=None` (raises); bad PUMA with `pumas` and touching geometries (donor used, row sums to 1).
-- Other helpers: `_calculate_utility_probabilities`, `_calculate_prior_distributions`, `_sample_utility_per_building` (determinism, gas only when `has_natgas_connection`, etc.).
+- `puma_id_series_for_join` (PUMACE10, GEOID, missing columns).
+- `zero_excluded_gas_utilities_and_renormalize`: no excluded cols (unchanged); zero + renormalize; bad PUMA with `pumas=None` (raises); bad PUMA with `pumas` and touching geometries (donor used, row sums to 1).
+- Other helpers: `calculate_utility_probabilities`, `calculate_prior_distributions`, `sample_utility_per_building` (determinism, gas only when `has_natgas_connection`, etc.).
