+++ data/resstock/utility/assign_utility_ny.py
@@ -1,18 +1,14 @@
-"""
-Utility assignment for ResStock buildings (NY).
+"""Utility assignment for ResStock buildings (NY).
 
-Equivalent to utils/utility_assignment_ny.R.
-Moved from utils/bsf to data/resstock.
-For create_hh_utilities: requires geopandas; building data from S3 or local parquet.
-For assign_utility_ny: requires S3Path.
+Thin wrapper around generic utility assignment functions in
+``data.resstock.utility.utils``, providing NY-specific configuration
+(utility name crosswalk, excluded gas utilities).
 """
 
 import argparse
 from typing import cast
 
 import geopandas as gpd
-import numpy as np
-import pandas as pd
 import polars as pl
 from cloudpathlib import S3Path
 from pygris import pumas as get_pumas
@@ -20,6 +16,10 @@ from pygris import pumas as get_pumas
 from data.resstock.utils import (
     load_state_configs,
     select_puma_and_heating_fuel_metadata,
+)
+from data.resstock.utility.utils import (
+    create_hh_utilities,
+    read_csv_to_gdf_from_s3,
 )
 from utils import get_aws_region
 from utils.utility_codes import get_ny_open_data_to_std_name
@@ -29,893 +29,6 @@ STORAGE_OPTIONS = {"aws_region": get_aws_region()}
 EXCLUDED_GAS_UTILITIES: frozenset[str] = frozenset(
     load_state_configs()["NY"]["excluded_gas_utilities"]
 )
-
-
-########################################################
-# GIS Utility Mapping
-########################################################
-def create_hh_utilities(
-    puma_and_heating_fuel: pl.LazyFrame,
-    electric_polygons: gpd.GeoDataFrame,
-    gas_polygons: gpd.GeoDataFrame,
-    pumas: gpd.GeoDataFrame,
-    config: dict,
-) -> pl.LazyFrame:
-    """
-    Create a LazyFrame of households with their associated utilities.
-
-    Uses Census PUMAs and utility service territory polygons to compute
-    overlap, then assigns electric/gas utilities to ResStock buildings by
-    sampling from PUMA-level probabilities. Requires geopandas.
-
-    Args:
-        puma_and_heating_fuel: LazyFrame with bldg_id, puma, heating_fuel columns
-        electric_polygons: GeoDataFrame of electric utility service territories
-        gas_polygons: GeoDataFrame of gas utility service territories
-        pumas: GeoDataFrame of Census PUMAs
-        config: State config dictionary (see data/resstock/state_configs.yaml).
-            Must contain ``state_crs``.
-
-    Returns:
-        LazyFrame with bldg_id, sb.electric_utility, sb.gas_utility.
-    """
-    utility_name_map = pl.DataFrame(
-        [
-            {"state_name": k, "std_name": v}
-            for k, v in get_ny_open_data_to_std_name().items()
-        ]
-    ).lazy()
-
-    # Calculate overlap between PUMAs and utilities
-    puma_elec_overlap = _calculate_puma_utility_overlap(
-        pumas, electric_polygons, config["state_crs"]
-    )
-
-    puma_gas_overlap = _calculate_puma_utility_overlap(
-        pumas, gas_polygons, config["state_crs"]
-    )
-
-    # Electric utility probabilities
-    # Force all buildings to have an electric utility (filter out "none")
-    puma_elec_probs = _calculate_utility_probabilities(
-        puma_elec_overlap,
-        utility_name_map,
-        handle_municipal=True,
-        filter_none=True,
-    )
-
-    # Gas utility probabilities
-    # Allow buildings without gas utility (keep "none" for non-gas buildings)
-    puma_gas_probs = _calculate_utility_probabilities(
-        puma_gas_overlap,
-        utility_name_map,
-        handle_municipal=False,
-        filter_none=False,
-    )
-    # Zero prior probability for small gas utilities; renormalize. If a PUMA has no
-    # gas utilities left, use the nearest neighbor PUMA's distribution (requires pumas).
-    puma_gas_probs = _zero_small_gas_utilities_and_renormalize(
-        puma_gas_probs, pumas=pumas, puma_and_heating_fuel=puma_and_heating_fuel
-    )
-
-    # Calculate prior probability distributions
-    elec_prior_weighted, gas_prior_weighted = _calculate_prior_distributions(
-        puma_elec_probs, puma_gas_probs, puma_and_heating_fuel=puma_and_heating_fuel
-    )
-
-    # Assign electric and gas to bldg's
-    building_elec = _sample_utility_per_building(
-        puma_and_heating_fuel, puma_elec_probs, "sb.electric_utility"
-    )
-
-    building_gas = _sample_utility_per_building(
-        puma_and_heating_fuel,
-        puma_gas_probs,
-        "sb.gas_utility",
-        only_when_fuel="Natural Gas",
-    )
-
-    # Print comparison summary
-    _print_comparison_summary(
-        building_elec,
-        building_gas,
-        elec_prior_weighted,
-        gas_prior_weighted,
-        puma_and_heating_fuel=puma_and_heating_fuel,
-    )
-
-    building_utilities = building_elec.join(
-        building_gas.select("bldg_id", "sb.gas_utility"), on="bldg_id", how="left"
-    )
-
-    return building_utilities
-
-
-def read_csv_to_gdf_from_s3(
-    s3_path: S3Path,
-    utility_type: str,
-    state_crs: int,
-    geometry_col: str = "the_geom",
-    crs: str = "EPSG:4326",
-) -> gpd.GeoDataFrame:
-    """
-    Read a CSV file from S3 with WKT geometry and convert to GeoDataFrame.
-
-    Parameters
-    ----------
-    s3_path : S3Path
-        S3Path to the CSV file (e.g., S3Path('s3://bucket/path/file.csv'))
-    utility_type : str
-        'electric' or 'gas'
-    state_crs : int
-        EPSG code of the state-specific projected CRS used for accurate area
-        calculations (e.g. 2260 for NY State Plane).  Read from
-        ``state_configs.yaml`` by callers.
-    geometry_col : str, default 'the_geom'
-        Name of the column containing WKT geometry
-    crs : str, default 'EPSG:4326'
-        Coordinate reference system for the geometries
-
-    Returns
-    -------
-    gpd.GeoDataFrame
-        GeoDataFrame with geometry from the WKT column
-    """
-    # Read CSV directly from S3
-    df = pd.read_csv(str(s3_path), low_memory=False)
-
-    # Convert numeric columns that might have been read as strings
-    # Skip geometry column and known string columns (like COMP_FULL/utility names)
-    if utility_type == "electric":
-        string_columns = {"comp_full", "utility", "state_name", "std_name"}
-    elif utility_type == "gas":
-        string_columns = {"COMP_FULL", "utility", "state_name", "std_name"}
-    for col in df.columns:
-        if col != geometry_col and col not in string_columns:
-            try:
-                df[col] = pd.to_numeric(df[col], errors="coerce")
-            except (ValueError, TypeError):
-                pass  # Leave non-numeric columns as-is
-
-    # Ensure string columns stay as strings (in case they were inferred as numeric)
-    for col in string_columns:
-        if col in df.columns:
-            df[col] = df[col].astype(str)
-
-    # Convert to GeoDataFrame
-    gdf = gpd.GeoDataFrame(
-        df, geometry=gpd.GeoSeries.from_wkt(df[geometry_col]), crs=crs
-    )
-
-    gdf = gdf.to_crs(epsg=state_crs)
-    if utility_type == "electric":
-        gdf = cast(gpd.GeoDataFrame, gdf.rename(columns={"comp_full": "utility"}))
-    elif utility_type == "gas":
-        gdf = cast(gpd.GeoDataFrame, gdf.rename(columns={"COMP_FULL": "utility"}))
-
-    return gdf
-
-
-def _calculate_puma_utility_overlap(
-    pumas: gpd.GeoDataFrame,
-    utility_gdf: gpd.GeoDataFrame,
-    state_crs: int,
-) -> pl.LazyFrame:
-    """Calculate overlap between PUMAs and utility service territories.
-
-    Args:
-        pumas: GeoDataFrame of PUMAs
-        utility_gdf: GeoDataFrame of utility polygons
-        state_crs: State-specific CRS (EPSG code) for accurate area calculations
-
-    Returns:
-        Polars LazyFrame with puma_id, pct_overlap, and all utility columns
-    """
-    # Transform PUMAs to state CRS and calculate area
-    puma_overlap = pumas.to_crs(epsg=state_crs).copy()
-    puma_overlap["puma_area"] = puma_overlap.geometry.area
-
-    # Transform utility polygons to state CRS
-    utility_polygons_transformed = utility_gdf.to_crs(epsg=state_crs)
-
-    # Perform spatial intersection
-    puma_overlap = gpd.overlay(
-        puma_overlap, utility_polygons_transformed, how="intersection"
-    )
-
-    # Calculate overlap area and percentage
-    puma_overlap["overlap_area"] = puma_overlap.geometry.area
-    puma_overlap["pct_overlap"] = (
-        puma_overlap["overlap_area"] / puma_overlap["puma_area"] * 100
-    ).astype(float)
-
-    # Drop geometry and select relevant columns
-    puma_overlap = puma_overlap.drop(columns=["geometry"])
-    utility_cols = [col for col in puma_overlap.columns if "utility" in col.lower()]
-    puma_overlap = puma_overlap[["PUMACE10", "pct_overlap"] + utility_cols].rename(
-        columns={"PUMACE10": "puma_id"}
-    )
-
-    # Convert to polars LazyFrame (utility column should already be string from source CSV)
-    return pl.from_pandas(puma_overlap).lazy()
-
-
-def _calculate_prior_distributions(
-    puma_elec_probs: pl.LazyFrame,
-    puma_gas_probs: pl.LazyFrame,
-    puma_and_heating_fuel: pl.LazyFrame,
-) -> tuple[dict[str, float], dict[str, float]]:
-    """Calculate prior probability distributions for electric and gas utilities.
-
-    Args:
-        puma_elec_probs: Wide-format LazyFrame with puma_id and electric utility probability columns
-        puma_gas_probs: Wide-format LazyFrame with puma_id and gas utility probability columns
-        puma_and_heating_fuel: LazyFrame with bldg_id, puma, heating_fuel
-
-    Returns:
-        Tuple of (elec_prior_weighted dict, gas_prior_weighted dict) for later comparison
-    """
-    # Electric: Calculate weighted average probability across all PUMAs
-    # (weighted by number of buildings in each PUMA)
-    puma_counts = (
-        puma_and_heating_fuel.group_by("puma").agg(pl.len().alias("count")).collect()
-    )
-    puma_elec_probs_df = cast(pl.DataFrame, puma_elec_probs.collect())
-    elec_prior = cast(
-        pl.DataFrame,
-        (
-            pl.LazyFrame(puma_elec_probs_df)
-            .join(
-                pl.LazyFrame(puma_counts),
-                left_on="puma_id",
-                right_on="puma",
-                how="left",
-            )
-            .with_columns(pl.col("count").fill_null(0))
-            .collect()
-        ),
-    )
-    utility_cols_elec = [c for c in puma_elec_probs_df.columns if c != "puma_id"]
-    elec_prior_weighted = {}
-    total_bldgs = elec_prior["count"].sum()
-    for util in utility_cols_elec:
-        weighted_prob = (
-            (elec_prior[util] * elec_prior["count"]).sum() / total_bldgs
-            if total_bldgs > 0
-            else 0
-        )
-        if weighted_prob > 0:
-            elec_prior_weighted[util] = weighted_prob
-
-    # Gas: Calculate weighted average probability (only for Natural Gas buildings)
-    gas_bldgs = puma_and_heating_fuel.filter(pl.col("has_natgas_connection"))
-    gas_puma_counts = gas_bldgs.group_by("puma").agg(pl.len().alias("count")).collect()
-    puma_gas_probs_df = cast(pl.DataFrame, puma_gas_probs.collect())
-    gas_prior = cast(
-        pl.DataFrame,
-        (
-            pl.LazyFrame(puma_gas_probs_df)
-            .join(
-                pl.LazyFrame(gas_puma_counts),
-                left_on="puma_id",
-                right_on="puma",
-                how="left",
-            )
-            .with_columns(pl.col("count").fill_null(0))
-            .collect()
-        ),
-    )
-    utility_cols_gas = [c for c in puma_gas_probs_df.columns if c != "puma_id"]
-    gas_prior_weighted = {}
-    total_gas_bldgs = gas_prior["count"].sum()
-    for util in utility_cols_gas:
-        weighted_prob = (
-            (gas_prior[util] * gas_prior["count"]).sum() / total_gas_bldgs
-            if total_gas_bldgs > 0
-            else 0
-        )
-        if weighted_prob > 0:
-            gas_prior_weighted[util] = weighted_prob
-
-    return elec_prior_weighted, gas_prior_weighted
-
-
-def _print_comparison_summary(
-    building_elec: pl.LazyFrame,
-    building_gas: pl.LazyFrame,
-    elec_prior_weighted: dict[str, float],
-    gas_prior_weighted: dict[str, float],
-    puma_and_heating_fuel: pl.LazyFrame,
-) -> None:
-    """Print comparison summary showing prior vs posterior distributions in a formatted table.
-
-    This summary compares the expected probability distribution (prior) with the actual
-    assignment distribution (posterior) to verify that sampling matches the intended probabilities.
-
-    Args:
-        building_elec: LazyFrame with bldg_id, sb.electric_utility
-        building_gas: LazyFrame with bldg_id, sb.gas_utility
-        elec_prior_weighted: Dictionary mapping electric utility names to prior probabilities
-        gas_prior_weighted: Dictionary mapping gas utility names to prior probabilities
-        puma_and_heating_fuel: LazyFrame with bldg_id, puma, heating_fuel (for filtering by fuel type)
-    """
-
-    def _print_comparison_table(
-        utility_type: str,
-        building_df: pl.LazyFrame,
-        utility_col: str,
-        prior_weighted: dict[str, float],
-        filter_fuel_type: str | None = None,
-    ) -> None:
-        """Print a formatted comparison table for one utility type.
-
-        Args:
-            filter_fuel_type: If provided, only compare buildings with this heating_fuel type
-        """
-        # Filter buildings by fuel type if specified (for gas utilities)
-        if filter_fuel_type is not None:
-            if filter_fuel_type == "Natural Gas":
-                building_df_filtered = building_df.join(
-                    puma_and_heating_fuel.select("bldg_id", "has_natgas_connection"),
-                    on="bldg_id",
-                    how="left",
-                ).filter(pl.col("has_natgas_connection"))
-            else:
-                building_df_filtered = building_df.join(
-                    puma_and_heating_fuel.select("bldg_id", "heating_fuel"),
-                    on="bldg_id",
-                    how="left",
-                ).filter(pl.col("heating_fuel") == filter_fuel_type)
-        else:
-            building_df_filtered = building_df
-
-        # Collect once and reuse for posterior, counts, and assignment stats
-        building_filtered_df = cast(pl.DataFrame, building_df_filtered.collect())
-        posterior_df = (
-            building_filtered_df.group_by(utility_col)
-            .agg(pl.len().alias("count"))
-            .with_columns((pl.col("count") / pl.col("count").sum()).alias("proportion"))
-            .sort("proportion", descending=True)
-        )
-
-        # Build comparison data (include None/null utilities)
-        all_utils = sorted(
-            set(
-                list(prior_weighted.keys())
-                + [r[utility_col] for r in posterior_df.iter_rows(named=True)]
-            ),
-            key=lambda x: (x is None, str(x) if x is not None else ""),
-        )
-
-        total_buildings = building_filtered_df.height
-        assigned_buildings = building_filtered_df.filter(
-            pl.col(utility_col).is_not_null()
-        ).height
-        unassigned_buildings = building_filtered_df.filter(
-            pl.col(utility_col).is_null()
-        ).height
-
-        if filter_fuel_type is not None:
-            print(
-                f"\nNote: Comparing only {filter_fuel_type} buildings "
-                f"({assigned_buildings} assigned, {unassigned_buildings} unassigned out of {total_buildings} total)"
-            )
-        else:
-            # For electric utilities, show assignment statistics
-            print(
-                f"\nAssignment Statistics: {assigned_buildings} assigned, "
-                f"{unassigned_buildings} unassigned (null) out of {total_buildings} total buildings"
-            )
-
-        comparisons = []
-        differences = []
-        for util in all_utils:
-            # Handle None/null utilities (they won't be in prior_weighted)
-            prior_prob = prior_weighted.get(util, 0.0) if util is not None else 0.0
-            prior_pct = prior_prob * 100
-            expected_count = prior_prob * total_buildings
-
-            # Find matching row (handle None comparison carefully)
-            post_row = None
-            for r in posterior_df.iter_rows(named=True):
-                if util is None and r[utility_col] is None:
-                    post_row = r
-                    break
-                elif util is not None and r[utility_col] == util:
-                    post_row = r
-                    break
-
-            posterior_pct = post_row["proportion"] * 100 if post_row else 0.0
-            actual_count = post_row["count"] if post_row else 0
-
-            # Format utility name for display
-            util_display = "None (unassigned)" if util is None else util
-            diff_pct = posterior_pct - prior_pct
-            diff_count = actual_count - expected_count
-            differences.append(abs(diff_pct))
-            comparisons.append(
-                {
-                    "utility": util_display,
-                    "prior_pct": prior_pct,
-                    "posterior_pct": posterior_pct,
-                    "diff_pct": diff_pct,
-                    "expected_count": expected_count,
-                    "actual_count": actual_count,
-                    "diff_count": diff_count,
-                }
-            )
-
-        # Print percentage comparison table
-        print(
-            f"\n{utility_type} Utilities - Prior vs Posterior Comparison (Percentages):"
-        )
-        print("This table compares the expected probability distribution (prior) with")
-        print(
-            "the actual assignment distribution (posterior) to verify sampling accuracy.\n"
-        )
-
-        # Table header
-        print(
-            f"{'Utility':<30} {'Prior %':>12} {'Posterior %':>14} {'Difference %':>15}"
-        )
-        print("-" * 73)
-
-        # Table rows
-        for comp in comparisons:
-            diff_str = f"{comp['diff_pct']:+.2f}"
-            print(
-                f"{comp['utility']:<30} "
-                f"{comp['prior_pct']:>11.2f}% "
-                f"{comp['posterior_pct']:>13.2f}% "
-                f"{diff_str:>14}"
-            )
-
-        # Print building count comparison table
-        print(f"\n{utility_type} Utilities - Expected vs Actual Building Counts:")
-        print(f"Total buildings: {total_buildings}")
-        print(f"{'Utility':<30} {'Expected':>12} {'Actual':>12} {'Difference':>15}")
-        print("-" * 73)
-
-        # Table rows
-        for comp in comparisons:
-            diff_str = f"{comp['diff_count']:+.0f}"
-            print(
-                f"{comp['utility']:<30} "
-                f"{comp['expected_count']:>11.0f} "
-                f"{comp['actual_count']:>11.0f} "
-                f"{diff_str:>14}"
-            )
-
-        # Statistics
-        if differences:
-            max_diff = max(differences)
-            avg_diff = sum(differences) / len(differences)
-            # Calculate standard deviation
-            mean_diff = avg_diff
-            variance = sum((d - mean_diff) ** 2 for d in differences) / len(differences)
-            std_diff = variance**0.5
-
-            print("\nDifference Statistics (Percentages):")
-            print(f"  Maximum absolute difference: {max_diff:.2f}%")
-            print(f"  Average absolute difference: {avg_diff:.2f}%")
-            print(f"  Standard deviation of differences: {std_diff:.2f}%")
-            if max_diff > 5.0:
-                print(
-                    "\n  WARNING: Maximum absolute difference exceeds 5%. "
-                    "Prior and posterior distributions may not match well."
-                )
-
-    print("\n" + "=" * 80)
-    print("PRIOR vs POSTERIOR COMPARISON SUMMARY")
-    print("=" * 80)
-
-    _print_comparison_table(
-        "Electric", building_elec, "sb.electric_utility", elec_prior_weighted
-    )
-
-    _print_comparison_table(
-        "Gas",
-        building_gas,
-        "sb.gas_utility",
-        gas_prior_weighted,
-        filter_fuel_type="Natural Gas",
-    )
-
-    print("\n" + "=" * 80 + "\n")
-
-
-def _calculate_utility_probabilities(
-    puma_overlap: pl.LazyFrame,
-    utility_name_map: pl.LazyFrame,
-    handle_municipal: bool = True,
-    filter_none: bool = False,
-    include_municipal: bool = False,
-) -> pl.LazyFrame:
-    """Calculate utility probabilities for each PUMA based on overlap percentages.
-
-    Args:
-        puma_overlap: LazyFrame with puma_id, utility, pct_overlap columns
-        utility_name_map: LazyFrame mapping from state_name to std_name for utilities
-        handle_municipal: Whether to transform "Municipal Utility:" names to "muni-" format
-        filter_none: Whether to filter out utilities named "none"
-        include_municipal: Whether to include municipal utilities in the utility assignment
-
-    Returns:
-        Wide-format LazyFrame with puma_id and probability columns for each utility
-    """
-    # Join with utility name map
-    probs = puma_overlap.join(
-        utility_name_map, left_on="utility", right_on="state_name", how="left"
-    )
-    probs = probs.with_columns(
-        pl.col("std_name").fill_null(pl.col("utility")).alias("utility")
-    ).drop("std_name")
-
-    # Handle municipal utilities if requested
-    if handle_municipal:
-        probs = probs.with_columns(
-            pl.when(pl.col("utility").str.starts_with("Municipal Utility:"))
-            .then(
-                pl.concat_str(
-                    [
-                        pl.lit("muni-"),
-                        pl.col("utility")
-                        .str.replace("Municipal Utility:", "")
-                        .str.strip_chars()
-                        .str.to_lowercase(),
-                    ]
-                )
-            )
-            .otherwise(pl.col("utility"))
-            .alias("utility")
-        )
-
-    # Aggregate overlap for duplicate (puma_id, utility) pairs
-    # (can happen when overlay produces multiple intersection fragments for non-contiguous territories)
-    probs = probs.group_by(["puma_id", "utility"]).agg(pl.col("pct_overlap").sum())
-
-    # Calculate probability as percentage of overlap within each PUMA
-    probs = probs.with_columns(
-        (pl.col("pct_overlap") / pl.col("pct_overlap").sum().over("puma_id")).alias(
-            "probability"
-        )
-    )
-
-    # Select relevant columns
-    probs = probs.select(["puma_id", "utility", "probability"])
-
-    # Exclude municipal (muni-*) utilities from probabilities if requested
-    if not include_municipal:
-        probs = probs.filter(~pl.col("utility").str.contains("muni-"))
-
-    # Filter out "none" utilities if requested
-    if filter_none:
-        probs = probs.filter(pl.col("utility") != "none")
-
-    # Pivot to wide format (pivot must be done on collected DataFrame)
-    probs_collected = cast(pl.DataFrame, probs.collect())
-    probs_pivoted = probs_collected.pivot(
-        index="puma_id", on="utility", values="probability", aggregate_function=None
-    )
-    probs = probs_pivoted.fill_null(0).lazy()
-
-    return probs
-
-
-def _puma_id_series_for_join(pumas: gpd.GeoDataFrame) -> pd.Series | None:
-    """Return a Series of puma_id values that match assign_utility_ny puma_id (e.g. 5-char)."""
-    if "PUMACE10" in pumas.columns:
-        return pumas["PUMACE10"].astype(str).str.zfill(5)
-    if "GEOID" in pumas.columns:
-        return pumas["GEOID"].astype(str).str[-5:]
-    return None
-
-
-def _zero_small_gas_utilities_and_renormalize(
-    puma_gas_probs: pl.LazyFrame,
-    pumas: gpd.GeoDataFrame | None = None,
-    puma_and_heating_fuel: pl.LazyFrame | None = None,
-) -> pl.LazyFrame:
-    """Set prior probability to zero for excluded gas utilities; renormalize.
-
-    For each PUMA row, columns corresponding to EXCLUDED_GAS_UTILITIES are set
-    to 0, then the row is renormalized so probabilities sum to 1. If any PUMA
-    would have all gas utilities zeroed (no gas utility left with positive
-    probability), either raises ValueError (when pumas is None) or uses the
-    nearest neighbor PUMA's gas probability distribution (when pumas is
-    provided). When puma_and_heating_fuel is provided, prints how many gas
-    buildings (has_natgas_connection) are in each affected PUMA.
-    """
-    gas_probs_df = cast(pl.DataFrame, puma_gas_probs.collect())
-    utility_cols = [c for c in gas_probs_df.columns if c != "puma_id"]
-    excluded_cols = [c for c in utility_cols if c in EXCLUDED_GAS_UTILITIES]
-
-    if not excluded_cols:
-        return puma_gas_probs
-
-    # Keep a copy of probabilities before zeroing (for debug output)
-    gas_probs_df_before_zero = gas_probs_df.clone()
-
-    # Zero out excluded gas utility columns
-    gas_probs_df = gas_probs_df.with_columns(
-        [pl.lit(0.0).alias(c) for c in excluded_cols]
-    )
-
-    # Row-wise sum of probabilities (all utility columns)
-    row_sums = gas_probs_df.select(pl.sum_horizontal(utility_cols)).to_series()
-    gas_probs_df = gas_probs_df.with_columns(row_sums.alias("_row_sum"))
-    bad_mask = row_sums == 0
-
-    if bad_mask.any():
-        bad_puma_ids = gas_probs_df.filter(bad_mask)["puma_id"].to_list()
-        if pumas is None:
-            raise ValueError(
-                "Zero gas probability for all utilities in PUMA(s) after excluding "
-                f"excluded gas utilities {sorted(EXCLUDED_GAS_UTILITIES)}. Affected puma_id(s): {bad_puma_ids}. "
-                "Excluding these utilities leaves no gas utility with positive probability."
-            )
-
-        # Building counts per PUMA (gas-only) for debug output
-        gas_bldg_counts: dict[str, int] = {}
-        if puma_and_heating_fuel is not None:
-            counts_df = cast(
-                pl.DataFrame,
-                puma_and_heating_fuel.filter(pl.col("has_natgas_connection"))
-                .group_by("puma")
-                .agg(pl.len().alias("n_bldgs"))
-                .collect(),
-            )
-            for row in counts_df.iter_rows(named=True):
-                puma_val = row["puma"]
-                key = str(puma_val).zfill(5) if puma_val is not None else ""
-                gas_bldg_counts[key] = int(row["n_bldgs"])
-
-        # Resolve PUMA id column in pumas for matching and geometry
-        puma_id_in_pumas = _puma_id_series_for_join(pumas)
-        if puma_id_in_pumas is None:
-            raise ValueError(
-                "pumas GeoDataFrame has no PUMACE10 or GEOID column; cannot find nearest neighbor."
-            )
-        pumas = pumas.copy()
-        pumas["_puma_id"] = puma_id_in_pumas.values
-
-        good_df = gas_probs_df.filter(~bad_mask)
-        good_puma_ids = good_df["puma_id"].to_list()
-        if not good_puma_ids:
-            raise ValueError(
-                "No PUMA has non-zero gas probability after excluding small utilities; "
-                "cannot assign nearest neighbor."
-            )
-
-        # Use full geometry for adjacency; centroids only as fallback/tiebreaker
-        pumas_geom = pumas.set_geometry("geometry")
-        pumas_geom = pumas_geom[pumas_geom.geometry.notna()].copy()
-        centroids = pumas_geom.geometry.centroid
-
-        fixed_rows = []
-        for bad_puma_id in bad_puma_ids:
-            bad_str = str(bad_puma_id).zfill(5)
-            bad_idx = pumas_geom["_puma_id"].astype(str) == bad_str
-            if not bad_idx.any():
-                raise ValueError(
-                    f"Bad PUMA id {bad_puma_id!r} not found in pumas GeoDataFrame."
-                )
-            bad_geom = pumas_geom.loc[bad_idx, "geometry"].iloc[0]
-            bad_centroid = centroids[bad_idx].iloc[0]
-
-            # Prefer PUMAs that touch the bad PUMA (adjacent); fall back to nearest centroid
-            adjacent_good: list[str] = []
-            for good_puma_id in good_puma_ids:
-                good_str = str(good_puma_id).zfill(5)
-                good_idx = pumas_geom["_puma_id"].astype(str) == good_str
-                if not good_idx.any():
-                    continue
-                good_geom = pumas_geom.loc[good_idx, "geometry"].iloc[0]
-                if bad_geom.touches(good_geom):
-                    adjacent_good.append(good_str)
-
-            if adjacent_good:
-                # Among adjacent PUMAs, pick nearest by centroid (tiebreaker)
-                best_donor = None
-                best_dist = float("inf")
-                for good_str in adjacent_good:
-                    good_idx = pumas_geom["_puma_id"].astype(str) == good_str
-                    good_centroid = centroids[good_idx].iloc[0]
-                    d = bad_centroid.distance(good_centroid)
-                    if d < best_dist:
-                        best_dist = d
-                        best_donor = good_str
-                nn_note = "adjacent (touching boundary)"
-            else:
-                # No adjacent good PUMA; fall back to nearest centroid
-                best_donor = None
-                best_dist = float("inf")
-                for good_puma_id in good_puma_ids:
-                    good_str = str(good_puma_id).zfill(5)
-                    good_idx = pumas_geom["_puma_id"].astype(str) == good_str
-                    if not good_idx.any():
-                        continue
-                    good_centroid = centroids[good_idx].iloc[0]
-                    d = bad_centroid.distance(good_centroid)
-                    if d < best_dist:
-                        best_dist = d
-                        best_donor = good_str
-                nn_note = (
-                    "no adjacent PUMA with gas; using nearest by centroid (fallback)"
-                )
-
-            if best_donor is None:
-                raise ValueError(f"No donor PUMA found for bad PUMA {bad_puma_id!r}.")
-            donor_row = good_df.filter(
-                pl.col("puma_id").cast(pl.Utf8).str.zfill(5) == best_donor
-            )
-            if donor_row.height == 0:
-                donor_row = good_df.filter(
-                    pl.col("puma_id").cast(pl.Utf8) == best_donor
-                )
-            if donor_row.height == 0:
-                donor_row = good_df.filter(pl.col("puma_id") == best_donor)
-            if donor_row.height == 0:
-                raise ValueError(f"Donor PUMA {best_donor!r} not found in good_df.")
-            # Copy donor's utility columns into a row for this bad puma_id
-            donor_vals = donor_row.select(utility_cols).row(0)
-            fixed_rows.append(
-                pl.DataFrame(
-                    {
-                        "puma_id": [bad_puma_id],
-                        **{c: [donor_vals[i]] for i, c in enumerate(utility_cols)},
-                        "_row_sum": [donor_row["_row_sum"][0]],
-                    }
-                )
-            )
-
-            # Debug: original row (before zeroing) for this PUMA
-            orig_row_df = gas_probs_df_before_zero.filter(
-                pl.col("puma_id").cast(pl.Utf8).str.zfill(5) == bad_str
-            )
-            if orig_row_df.height == 0:
-                orig_row_df = gas_probs_df_before_zero.filter(
-                    pl.col("puma_id").cast(pl.Utf8) == bad_str
-                )
-            if orig_row_df.height == 0:
-                orig_row_df = gas_probs_df_before_zero.filter(
-                    pl.col("puma_id") == bad_puma_id
-                )
-            orig_vals = (
-                orig_row_df.select(utility_cols).row(0) if orig_row_df.height else None
-            )
-            orig_probs = (
-                {utility_cols[i]: orig_vals[i] for i in range(len(utility_cols))}
-                if orig_vals
-                else {}
-            )
-            excluded_that_had_prob = [
-                c for c in excluded_cols if orig_probs.get(c, 0.0) > 0
-            ]
-            small_probs_before = {c: orig_probs[c] for c in excluded_that_had_prob}
-            donor_probs = {
-                utility_cols[i]: donor_vals[i] for i in range(len(utility_cols))
-            }
-            prior_before_str = ", ".join(
-                f"{u}={orig_probs.get(u, 0):.3f}"
-                for u in utility_cols
-                if orig_probs.get(u, 0) > 0
-            )
-            after_nn_str = ", ".join(
-                f"{u}={donor_probs.get(u, 0):.3f}"
-                for u in utility_cols
-                if donor_probs.get(u, 0) > 0
-            )
-            row_sum_before = sum(orig_probs.get(u, 0) for u in utility_cols)
-
-            n_bldgs = gas_bldg_counts.get(bad_str, 0)
-            print(
-                f"[assign_utility_ny] PUMA {bad_puma_id!r} had zero gas probability "
-                f"after excluding small utilities; using donor PUMA {best_donor!r} "
-                f"({nn_note}; distance={best_dist:.0f} m). "
-                f"Affected bldg_ids (gas buildings in this PUMA): {n_bldgs}."
-            )
-            print(
-                f"  Excluded gas utilities zeroed in this PUMA: {excluded_that_had_prob}; "
-                f"their prior probs before removal: {small_probs_before}."
-            )
-            print(
-                f"  Prior (before removal): {prior_before_str} (row sum={row_sum_before:.3f})"
-            )
-            if set(excluded_that_had_prob) == set(
-                u for u in utility_cols if orig_probs.get(u, 0) > 0
-            ):
-                print(
-                    "  → So before the fix, all gas bldg_ids in this PUMA would have been "
-                    f"assigned to one of {excluded_that_had_prob}."
-                )
-            print(f"  After nearest-neighbor approximation: {after_nn_str}")
-
-        gas_probs_df = pl.concat([good_df, pl.concat(fixed_rows)])
-
-    gas_probs_df = gas_probs_df.drop("_row_sum")
-
-    # Renormalize so each row sums to 1
-    row_sums = gas_probs_df.select(pl.sum_horizontal(utility_cols)).to_series()
-    gas_probs_df = gas_probs_df.with_columns(
-        [(pl.col(c) / row_sums).alias(c) for c in utility_cols]
-    )
-    return gas_probs_df.lazy()
-
-
-def _sample_utility_per_building(
-    bldgs: pl.LazyFrame,
-    puma_probs: pl.LazyFrame,
-    utility_col_name: str,
-    only_when_fuel: str | None = None,
-    seed: int = 42,
-) -> pl.LazyFrame:
-    """For each building, sample one utility from its PUMA's probability distribution.
-
-    Args:
-        bldgs: LazyFrame with bldg_id, puma, heating_fuel
-        puma_probs: Wide-format LazyFrame with puma_id and probability columns for each utility
-        utility_col_name: Name for the output utility column
-        only_when_fuel: If provided, only sample when heating_fuel matches this value.
-            If only_when_fuel == "Natural Gas", assign all bldg_ids that have has_natgas_connection==True
-
-    Returns:
-        LazyFrame with bldg_id and the sampled utility column
-    """
-    # Join buildings with their PUMA probabilities
-    bldgs_joined = bldgs.join(
-        puma_probs, left_on="puma", right_on="puma_id", how="left"
-    )
-
-    # Get utility column names (all columns except puma_id from puma_probs)
-    # Sort for deterministic ordering
-    puma_probs_df = cast(pl.DataFrame, puma_probs.collect())
-    utility_cols = sorted([c for c in puma_probs_df.columns if c != "puma_id"])
-
-    # Convert to pandas for row-wise sampling with numpy (need to collect for pandas)
-    bldgs_joined_df = cast(pl.DataFrame, bldgs_joined.collect())
-    bldgs_pd = bldgs_joined_df.to_pandas()
-
-    # Sort by bldg_id for deterministic ordering
-    bldgs_pd = bldgs_pd.sort_values("bldg_id").reset_index(drop=True)
-
-    # Set random seed for deterministic sampling
-    np.random.seed(seed)
-
-    def sample_utility(row) -> str | None:
-        """Sample one utility based on probability distribution."""
-        if only_when_fuel is not None:
-            if only_when_fuel == "Natural Gas":
-                if not row["has_natgas_connection"]:
-                    return None
-            else:
-                if row["heating_fuel"] != only_when_fuel:
-                    return None
-
-        # Get probabilities for this row and convert to numeric (float)
-        probs = pd.to_numeric(row[utility_cols].values, errors="coerce").astype(float)
-
-        # Handle cases where all probabilities are 0 or NaN
-        if np.all(np.isnan(probs)) or np.sum(probs) == 0:
-            return None
-
-        # Replace NaN with 0 and normalize probabilities
-        probs = np.nan_to_num(probs, nan=0.0)
-        probs = probs / np.sum(probs)
-
-        # Sample one utility based on probabilities (replace=False is explicit but doesn't matter for size=1)
-        sampled_utility = np.random.choice(
-            utility_cols, size=1, replace=False, p=probs
-        )[0]
-        return sampled_utility
-
-    # Apply sampling to each row
-    bldgs_pd[utility_col_name] = bldgs_pd.apply(sample_utility, axis=1)
-
-    # Convert back to polars LazyFrame and select only needed columns
-    result = pl.from_pandas(bldgs_pd[["bldg_id", utility_col_name]]).lazy()
-
-    return result
 
 
 def assign_utility_ny(
@@ -943,12 +56,22 @@ def assign_utility_ny(
         LazyFrame with all original columns plus sb.electric_utility and sb.gas_utility
     """
     puma_and_heating_fuel = select_puma_and_heating_fuel_metadata(input_metadata)
+
+    utility_name_map = pl.DataFrame(
+        [
+            {"state_name": k, "std_name": v}
+            for k, v in get_ny_open_data_to_std_name().items()
+        ]
+    ).lazy()
+
     building_utilities = create_hh_utilities(
         puma_and_heating_fuel=puma_and_heating_fuel,
-        config=config,
         electric_polygons=electric_polygons,
         gas_polygons=gas_polygons,
         pumas=pumas,
+        utility_name_map=utility_name_map,
+        state_crs=config["state_crs"],
+        excluded_gas_utilities=EXCLUDED_GAS_UTILITIES,
     )
 
     # Drop existing utility columns if they exist (equivalent to ADD COLUMN IF NOT EXISTS)
@@ -1051,8 +174,6 @@ if __name__ == "__main__":
     gas_poly_dir = S3Path(args.gas_poly_dir)
     electric_poly_path = S3Path(electric_poly_dir / args.electric_poly_filename)
     gas_poly_path = S3Path(gas_poly_dir / args.gas_poly_filename)
-
-    from data.resstock.utils import load_state_configs
 
     ny_config = load_state_configs()["NY"]
